move test_grouped_dims to test/null
it's a pure helper
